refactor: SafetyError型をerrors.goに統合しContext機能を削除
セルフレビューの指摘に従い、以下の改善を実施:

1. ファイル名の不整合を修正
   - internal/stager/safety_errors.goを削除
   - SafetyError型をinternal/stager/errors.goに統合

2. 不要なContext機能を削除（YAGNI原則）
   - WithContext/GetContextメソッドを削除
   - Context mapフィールドを削除
   - 関連テストを削除

3. エラー型名の冗長性を修正
   - ErrorTypeStagingAreaNotClean → StagingAreaNotClean
   - 他のエラー型名も同様に簡潔化

既存のStagerErrorと同様のシンプルな設計に統一。
